[Windows] Convert build to MSYS2 from Cygwin
While we have switched from Cygwin to MSYS2 due to an upstream bug
linked below, we still need to figure out why there's a problem building
in this new setup.

Upstream issue that makes Cygwin unable to build FIPS-compliant module
according to the security policy:
- https://mta.openssl.org/pipermail/openssl-dev/2016-February/005392.html

While the fix is a one-liner, we cannot apply it to the source of the
canister without breaking the security policy.